pyproject.toml
[project]
name = "traffic-project-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite==0.20.0",
    "annotated-types==0.7.0",
    "anyio==4.6.2.post1",
    "asyncpg==0.30.0",
    "atlastk==0.13.3",
    "click==8.1.7",
    "fastapi==0.115.5",
    "h11==0.14.0",
    "idna==3.10",
    "iso8601==2.1.0",
    "pandas>=2.2.3",
    "psutil==6.1.0",
    "psycopg2-binary>=2.9.11",
    "pydantic==2.10.2",
    "pydantic-core==2.27.1",
    "pypika-tortoise==0.3.1",
    "python-dotenv==1.0.1",
    "pytz==2024.2",
    "scikit-learn>=1.6.1",
    "shapely==2.0.6",
    "sniffio==1.3.1",
    "starlette==0.41.3",
    "tensorflow>=2.19.0",
    "tqdm>=4.67.3",
    "typing-extensions==4.12.2",
    "uvicorn==0.32.1",
]
